:bug: Fixed push swap segfaults

+++ Core/PushSwap/src/stack/ft_stack_size.c
@@ -6,7 +6,7 @@
 /*   By: ibour <support@toujoustudios.net>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 10:59:56 by ibour             #+#    #+#             */
-/*   Updated: 2024/11/05 11:00:45 by ibour            ###   ########.fr       */
+/*   Updated: 2024/11/05 17:16:41 by ibour            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_stack_size(const t_stack *stack)
 {
-	size_t	size;
+	int	size;
 
 	size = 0;
 	while (stack) {
